build: Provide wrapper functions instead of aliases in DLL built with MSVC
On Windows we cannot use proper aliases, and using the /EXPORT linker
flag is too cumbersome, as it does not work when static linking, and
when dynamic linking it does not make the aliases visible within the
DLL itself.

Just use normal function wrapper in those cases, which are way more
maintainable.